backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
tensorflow==2.14.0
pillow==10.0.1
numpy==1.26.4
pandas==2.2.2
